@@ -1,23 +1,36 @@
 import { Routes } from '@angular/router';
 import { SettingsComponent, NotFound, ListComponent, PlayerComponent } from './ui';
-import { SignIn } from '../auth';
-
+import { SignIn } from '../auth/signIn.component';
+import { AuthGuard } from './auth.guard';
 export const routes: Routes = [
+  {
+    path: 'auth',
+    component: SignIn
+  },
   {
     path: 'list',
     component: ListComponent,
-    canActivate: [],
-    pathMatch: 'full'
+    canMatch: [AuthGuard]
+  },
+  {
+    path: 'list/:pageIndex',
+    component: ListComponent,
+    canMatch: [AuthGuard]
   },
   {
     path: 'player',
     component: PlayerComponent,
-    //canActivate: [MsalGuard]
+    canMatch: [AuthGuard]
   },
   {
     path: 'settings',
     component: SettingsComponent,
-    // canActivate: [MsalGuard]
+    canMatch: [AuthGuard]
+  },
+  {
+    path: '',
+    redirectTo: '/list',
+    pathMatch: 'full'
   },
   {
     path: '**',
