@@ -1,7 +1,10 @@
 
-
-export const AUTH_TOKEN: string = "auth_token";
-export const ACCESS_TOKEN: string = 'access_token';
+export const STATE: string = 'state';
 export const CODE_WITH_PKCE: string = 'code_with_pkce';
-export const EXPIRE_TIME: string = 'expire_time';
-
+export const CODE_WITH_PKCE_EXPIRE: string = 'code_with_pkce_expire';
+export const CODE_WITH_PKCE_EXPIRE_TIME = 1000 * 60 * 10;
+export const ACCESS_TOKEN: string = 'access_token';
+export const AUTH_TOKEN: string = "auth_token";
+export const AUTH_TOKEN_EXPIRE: string = 'auth_token_expire';
+export const REFRESH_TOKEN_EXPIRE = 'fresh_token_expire';
+export const AUTH_TOKEN_EXPIRE_TIME: number = 24*60*60*1000;
