@@ -1,6 +1,5 @@
 export interface ApiRequest {
     endpoint: string;
-    scopes: string[];
     forceRefresh?: boolean;
 }
 export namespace ApiRequest {
@@ -10,12 +9,9 @@ export namespace ApiRequest {
     //     forceRefresh: false
     // // }
     export const Authorize: ApiRequest = {
-      endpoint: `https://login.microsoftonline.com/55604d97-faf8-4a3c-8f8d-c7a4fbc9b8b6/oauth2/v2.0/authorize`,
-      scopes: ["User.Read"],
-      forceRefresh: false
+      endpoint: `https://login.microsoftonline.com/consumers/oauth2/v2.0/authorize`,
     }
     export const Token: ApiRequest = {
-      endpoint: `https://login.microsoftonline.com/55604d97-faf8-4a3c-8f8d-c7a4fbc9b8b6/oauth2/v2.0/token`,
-      scopes: ["user.read"]
+      endpoint: `https://login.microsoftonline.com/consumers/oauth2/v2.0/token`,
     }
 }
