@@ -1,23 +1,27 @@
 
-import { ApplicationInitStatus, Component, inject, OnInit } from "@angular/core";
-import { fromQueryParams } from "./utils";
-import { GetCodeWithPKCEResponse } from "./models";
-import { AccessTokenProvider } from "./token";
-import { CODE_WITH_PKCE } from "./Constant";
+import { Component, inject, OnInit } from "@angular/core";
+
 import { Router } from "@angular/router";
+import { GetAccessTokenOptions } from "./options";
+import { HttpClient } from "@angular/common/http";
+import { RequestAccessToken } from "./accessToken.auth";
+import { HandleCodeWithPKCE } from "./codeWithPKCE.auth";
 @Component({
   template: ``,
   standalone: true
 })
 export class SignIn implements OnInit {
-  private provider: AccessTokenProvider = inject(AccessTokenProvider);
+  private options: GetAccessTokenOptions = inject(GetAccessTokenOptions);
+  private http: HttpClient = inject(HttpClient);
   private router: Router = inject(Router);
   ngOnInit(): void {
-    if(location.search.indexOf('code') < 0) {
-      var res = fromQueryParams<GetCodeWithPKCEResponse>(location.search);
-      localStorage.setItem(CODE_WITH_PKCE, location.search);
-      this.provider.reload();
-    }
-    this.router.navigateByUrl('/');
+    if(HandleCodeWithPKCE())
+      RequestAccessToken(this.options, this.http)
+        .subscribe({
+          next: _ => this.router.navigate(['/']),
+          error: console.error
+        });
+    else
+      console.error(location.search.slice(1).replaceAll('&', '\n'));
   }
 }
