@@ -7,14 +7,19 @@ export interface ApiRequest {
 }
 
 export namespace ApiRequest {
+
     export const me: ApiRequest = {
         endpoint: `https://graph.microsoft.com/v1.0/me`,
-        scopes: ["User.Read"]
+        scopes: ["https://graph.microsoft.com/User.Read"]
     };
     export const files: ApiRequest = {
 
         endpoint: `https://graph.microsoft.com/v1.0/drives/B2D7A30C38920DE8/items/B2D7A30C38920DE8!134726/children?$expand=thumbnails`,
-        scopes: ["User.Read", "Files.Read", "Files.Read.All"],
+        scopes: [
+          "https://graph.microsoft.com/User.Read",
+          "https://graph.microsoft.com/Files.Read",
+          "https://graph.microsoft.com/Files.Read.All"
+        ],
         forceRefresh: true
     };
 }
