@@ -1,4 +1,8 @@
-
+export interface Thumbnail {
+  height: number;
+  url: string;
+  width: number;
+}
 
 export interface VideoMetadata {
   ["@microsoft.graph.downloadUrl"]: string
@@ -25,14 +29,41 @@ export interface VideoMetadata {
   };
   thumbnails: {
     id: string;
-    large: { height: number, url: string, width: number };
-    medium: { height: number, url: string, width: number };
-    small: { height: number, url: string, width: number };
+    large: Thumbnail;
+    medium: Thumbnail;
+    small: Thumbnail;
   }[];
 }
+
+
 export interface DriveResponse<T> {
   [`@odata.context`]: string;
   [`@odata.count`]: number;
   [`@odata.nextLink`]: string;
   value: T[];
+}
+
+
+export interface VideoSource {
+  src: string;
+  title: string;
+  thumb: {
+    grid: Thumbnail,
+    row: Thumbnail
+  }
+  size: number;
+  mimeType: string;
+}
+
+export function ToViduoSource(video: VideoMetadata): VideoSource {
+  return {
+    src: video["@microsoft.graph.downloadUrl"],
+    title: video.name,
+    thumb: {
+      grid: video.thumbnails[0].large,
+      row: video.thumbnails[0].small
+    },
+    size: video.size,
+    mimeType: video.file.mimeType
+  };
 }
